Fixed warnings

+++ Xlnt/Xlnt.cpp
@@ -3,6 +3,8 @@
 #include <Core/Core.h>
 #include "Xlnt.h"
 
+#include <Functions4U/EnableWarnings.h>
+
 namespace Upp {
 
 String GetCell(int col, int row) {
@@ -17,16 +19,16 @@ String GetCell(int col, int row) {
 
 void XlsxFill(xlnt::worksheet &ws, const Grid &g, bool removeEmpty) {
 	for (int c = 0; c < g.cols(); ++c)
-		ws.column_properties(c+1).width = g.GetWidth(c)/5.;
+		ws.column_properties(unsigned(c+1)).width = g.GetWidth(c)/5.;
 	
 	if (g.GetNumHeaderRows() > 0) 
 		ws.range(xlnt::range_reference(xlnt::cell_reference(1, 1),
-                 					   xlnt::cell_reference(g.cols(), g.GetNumHeaderRows()))).font(xlnt::font().bold(true))
+                 					   xlnt::cell_reference((unsigned)g.cols(), (unsigned)g.GetNumHeaderRows()))).font(xlnt::font().bold(true))
                  					   														 .fill(xlnt::fill::solid(xlnt::rgb_color(230, 230, 230)));
 	
 	if (g.GetNumHeaderCols() > 0) 
 		ws.range(xlnt::range_reference(xlnt::cell_reference(1, 1),
-                 					   xlnt::cell_reference(g.GetNumHeaderCols(), g.rows()))).font(xlnt::font().bold(true))
+                 					   xlnt::cell_reference((unsigned)g.GetNumHeaderCols(), (unsigned)g.rows()))).font(xlnt::font().bold(true))
                  					   														 .fill(xlnt::fill::solid(xlnt::rgb_color(230, 230, 230)));
 	
 	for (int r = 0, row = 0; r < g.rows(); ++r) {
@@ -44,13 +46,13 @@ void XlsxFill(xlnt::worksheet &ws, const Grid &g, bool removeEmpty) {
 			for (int c = 0; c < g.cols(); ++c) {
 				const Value &v = g.Get(r, c);
 				if (v.Is<bool>())
-					ws.cell(c+1, row+1).value(bool(v));
+					ws.cell((unsigned)(c+1), (unsigned)(row+1)).value(bool(v));
 				else if (v.Is<int>())
-					ws.cell(c+1, row+1).value(int(v));
+					ws.cell((unsigned)(c+1), (unsigned)(row+1)).value(int(v));
 				else if (v.Is<double>())
-					ws.cell(c+1, row+1).value(double(v));
+					ws.cell((unsigned)(c+1), (unsigned)(row+1)).value(double(v));
 				else 
-					ws.cell(c+1, row+1).value(~v.ToString());
+					ws.cell((unsigned)(c+1), (unsigned)(row+1)).value(~v.ToString());
 			}
 			row++;
 		}
